drivers: wlan: Cancel remain on channel without con mode
At the time of driver exit, driver set con mode to VOS_STA_MODE.
If Remain on channel is in progress, at the time of hdd_stop_adapter
driver do cancel remain on channel.
WLANSAP_CancelRemainOnChannel cancel remain on channel if
con mode is VOS_STA_SAP_MODE which is not valid any more.
Fix this issue by not constraining cancel remain on channel
on basis of con mode.

Change-Id: I0f14c92f24b0f1b38a46142175e0e6d31e8bb318
CRs-Fixed: 1072605