drivers: wlan: Send ADDBA req to test the connection status
In some cases AP silently kick outs the DUT and DUT is unaware
of its connection status. After this when DUT send packets to AP,
AP drops the packets.

To check if the connection is alive send ADD BA request to AP. In
this change ADDBA req is send on a particular tid for which BA
session does not exist.

Change-Id: Ie88a5c88b8b880cfc694a5899fc6be03bf52a2bc
CRs-fixed: 1116816