drivers: wlan: Process roam disable after roam completion
Currently roam disable is allowed using SETROAMMODE even if roaming
is in progress. This can cause mess in state machine as driver falls
back to LFR-1.0 and can result in crash. Add pending_roam_disable to
mark pending of roam disable when roaming is in progress and allow
preauth/reassoc to go if preauth/reassoc in progress.

With this fix, driver takes following actions.
1. If pending_roam_disable is pending after preauth failure,
   driver does not proceed for further preauth and disables roaming.
2. If pending_roam_disable is pending after reassoc completion,
   driver updates CSR with new AP information and disables roaming.

Change-Id: I9cfcbd21db046d399c2cdc78b1678d3773aea2cd
CRs-Fixed: 2045775